r/virtual_machine: Roll back clones on errors with config expansion
The interface for expandVirtualMachineConfigSpec has changed to
accommodate errors during config expansion (namely to check for invalid
vApp properties). However, as the expansion happens post-clone but
before the initial configuration, errors risk putting the Terraform
state in an irrecoverable situation should the configuration fail here.

Moved the rollback functionality that we have for the post-clone
reconfigure into a helper function so that we can leverage it here.

Also, changed the wording of the error message so that it's a bit
friendlier to the average user.