r/virtual_machine: Block vApp properties on from-scratch scenarios
This blocks vApp properties from being set on virtual machines created
from scratch. As this support is targeted towards VMs that haven been
cloned from a template that has been imported from OVF/OVA, we expect
the vApp configuration to exist (and further to that we don't support
setting unknown keys). This basically means that it's impossible to set
vApp properties on new virtual machines with this resource, or it should
be.

This commit corrects a few crash scenarios that are happening when
trying to use this block on new virtual machines as it stands. It also
blocks situations where someone tries to add configuration after the
fact or on a virtual machine where no vApp configuration exists, which
would be possible on both the from-scratch and clone paths.