r/virtual_machine: Fix tests
The tests were not working with our lab as the seemed to be expecting
DHCP for network configuration. Added in the appropriate configuration
options that are set up for the rest of the clone tests, with the
customization options just moved to their appropriate cloud-config
key/value settings. This may need to be changed if CoreOS ever drops
support for the old cloud-init guestinfo keys in favour of Ignition.