r/virtual_machine: Introduce label, path, and uuid for disks
This shifts the disk virtual device logic over to do the following:

* The configuration key for virtual disks is now "label" an arbitrary
name local to Terraform that only identifies the disk in configuration,
and allows it to freely move around in terms of location. Unlike "name",
which was the previous identifier, label does not influence the actual
VMDK filename.
* A new attribute, "path", has been added - this either 1) is a computed
attribute that gives the path to the disk for unattached disks, or 2)
defines the path for an external virtual disk when "attach" is set. This
is derived from "name" when attaching without setting path.
* UUID is now the source of truth once disks have been created and
committed to state. This replaces device address, which now is only the
fallback when no UUID exists in state yet, such as right after a VM is
created or a disk is added during an update operation. This helps to
account for if someone manually changes the position of a disk on the
SCSI bus, which can be a serious failure scenario if the new position is
one that is currently occupied by another disk in state.

This allows us to lift a number of storage vMotion restrictions and
possibly avoid any other issues that may arise from depending on file
name so much.

Tests are still pending, and we need to review the state migration code,
and also check to determine if a new schema version will be necessary to
ensure that short-name disks are matched up properly with long-name
disks that may have been committed to state before.