r/virtual_machine: Bump schema version, add state migration, more
* Looks like we will need to grab the disk UUIDs for this to work
properly on the next read operation that happens after upgrade. Hence
we've bumped the schema version, with the V2 -> V3 migration path
populating the UUIDs for all present disks.
* diskLabelOrName now truncates paths to their base. This is what we
have been doing anyway for lining up disks in state, so there should be
no risk of regressions. Path calculations are handled via another path,
diskPathOrName, which is unaffected by this.
* Restore the behaviour of carrying forward the name attribute during
diffs, but *only* if label has not been defined in the new diff.
* Made the deprecation notice for the name field a bit better to read by
making it a pre-formatted string, also added in a link to the label
attribute so that we can refer people to it when this goes live.
* A few other updates to remove some more dead code.