r/virtual_machine: Changes to import and V1 -> V2 migration behaviour
This also fixes import in general (it had not been checked yet and would
have more than likely been broken).

Import functionality has now significantly changed and will need to be
documented.

In addition to validating SCSI types and what not, We now partially
populate state on import. This is so that the next refresh (which
happens after an import) can properly populate the rest of the state for
a disk.

The following fields are populated:
* device_address: With the proper device address of the disk.
* key: With (disknum + 1) * -1. This is to ensure DiskRefreshOperation
sees the device as a new disk and will not attempt to match it with a
UUID in state.
* label:  With a generic label that needs to be specified in config for
the import to work properly. This cannot be enforced as the process does
not have access to config, however keep_on_remove is set to ensure that
a disk only gets detached on errors (although this could still lead to
service interruptions and data corruption, depending on the situation).
The convention is diskN, starting with disk0.

The V1 to V3 state migration leans on this import functionality, and
ultimately yields a V3 state with *no* UUID populated, unlike V2 -> V3.
This is mitigated by the fact that thanks to the import functionality,
the disks show up as newly created, which allows the next
DiskRefreshOperation call (which happens as part of the next refresh) to
pick up the UUID, along with the rest of the attributes that should make
it to state.