r/virtual_machine: Fix disk device deletion logic
This fixes a couple of issues.

It appears some changes in virtual disk management (quite a while ago)
had actually broken virtual device deletion, at the very least when
removing devices from the middle of a Terraform state (ie: disk 2 of 3).
The effect was that no virtual disks were actually being deleted when
they should have been. However, the current logic (which had been
updated as part of the transition to UUIDs and also due to this problem
in general) had an error in the deletion conditional and was deleting
*all* disks.

This has now been fixed to correct both the old and new issues.

The original error was due to a duplication of the logic that we use in
network interfaces, where removing a device in the middle of the stack
means that the new set will be smaller than the old one, with the last
device removed and the middle one reconfigured. This is not the case
with virtual disks - we do extra diff customization to provide extra
verbosity during the device management process to tell you exactly what
is happening to your disks. The new, shrunk set is only read in after
the process is completed, during the post-update read.

The question remains as to how this passed acceptance tests for so long.
I will be investigating the acceptance test a bit more - I already
debugged the check and it seems to be working, but something is
obviously amiss that is keeping it from functioning correctly.