r/virtual_machine: Ensure path is not copied during post-clone
During post-clone, we do a read-in and a diff of all cloned virtual
disks to check to make sure they don't need changing (ie: growing
disks that aren't linked clones). During this situation, some keys are
ignored as they will always be empty or contain bad data that needs to
be populated after committing the diff to state post-apply.

"path" was a missed field here, and it has been added now. We don't
bother skipping this if attach is set to true, as since this is during
the evaluation of the template's disks specifically, there is no chance
these disks will be externally attached.