r/virtual_machine: Fix multiple change operations to a single device
Due to how the DeviceChange struct works in API, it's possible to end up
setting up a list of DeviceChange operations that possibly point to
the same underlying virtual device. In this instance, with our logic,
any subsequent operations that result in a device change will end up
changing all operations before it into the same state as the most
current version of that virtual device, causing an invalid operation.

This can be seen in a scenario that a VM is cloned from template, with
both the SCSI bus type and disk size changing at the same time:

* The SCSI type change will add an operation that changes the controller
the disk is attached to.
* The second operation will change the size of the virtual disk.

Both of these operations will operate on the same VirtualDisk device
instance and if the structure is not duplicated, this will result in the
first device change operation being identical to the second, causing the
subsequent reconfigure operation to fail.