Move end-to-end example to more general location, and update
This should be merged with a pending post-1.0 blog article update that
will be made to ensure that the article reflects the disk addressing as
it stands in 1.3.

The update (moving from name to label to address disks in VM resources)
is a significant enough update that I want to make sure that the example
is updated, as it received a significant amount of press when it was
released and hence serves as a good "getting started" reference example,
and I don't want people to use it only to find out that a key field in
it has been deprecated already.