Add OVF deployment example
This example demonstrates how one can use the vApp property management
features of the vsphere_virtual_machine resource to deploy a 3-node
webapp deployment using CoreOS, Ignition, and provisioners. A README is
provided for documentation.

The web demonstration application used, http-hello-go, is one I use to
demonstrate similar patterns in other projects, and can be found at the
link below. The code does not self-contain the app to avoid possible
licensing issues, but rather downloads and builds it as part of the
orchestrated build process.

https://github.com/vancluever/http-hello-go