r/virtual_machine: Switch ContainerView search to use manual filter
There seems like there is possibly a bug in RetrieveWithFilter when it
comes to MO's with missing keys and the function returning a false
positive for the object in question. This has caused issues with the VM
search functionality where duplicate UUID errors will be returned for
VMs with UUIDs that are otherwise unique within vSphere inventory if
there are orphaned VMs in inventory that lack a configuration (as in the
whole config data object is missing).

This bug applies to the provider when being used with vSphere 6.0 and
earlier.