r/virtual_machine: Use CapacityInKB when CapacityInBytes unavailable
Under certain situations, namely vCenter upgrades, the CapacityInBytes
field for virtual disks gets cleared,  which will cause issues fetching
settings for virtual disks on templates and virtual machines
post-upgrade.

This updates the virtual disk logic for both resources and data sources
so that if the value in CapacityInBytes is missing (read: less than 1),
we fall back to CapacityInKB, which will still be populated.