r/virtual_machine: Don't commit empty disk sharing value to state
This is another post-upgrade thing. While vCenter will be upgraded, VM
hardware versions may remain at specific versions, especially if the
underlying hypervisors have not been upgraded yet. In this situation,
disk sharing will be an unset value and setting it will yield an error,
hence we skip setting it so that it does not create a diff.

This kind of highlights the need for us to be aware of the VM hardware
version in the future, in addition to the specific version of vCenter
the endpoint is on.