r/virtual_machine: Handle missing environment browser
The virtual machine resource relies on the cluster or standalone host
it's deployed on for information regarding a default hardware set or
guest OS family information. This is obtained via the EnvironmentBrowser
MO referenced by the respective cluster or host's ComputeResource MO.

There are certain situations where this attribute will be missing, two
of which have been identified so far via bug reports:

* A cluster without hosts will not contain an environment browser.
* vSphere licensing issues may cause a missing environment browser (#409
reported an expired vCenter).

This update wraps our fetching of the environment browser by the two
current compute resource helpers that utilize it in a helper that
validates that we actually have an environment browser before proceeding
to load it up, issuing an error alluding to the above if it is missing.

Fixes #411.
May also correct #409.