provider: Start work to allow persistence of SOAP and REST sessions
This commit contains most of the SOAP session persistence code.

Persisting sessions to disk will work mostly in the same way that govc
does, and in fact most of the code here has been adapted from this work.
However, we also need to persist REST sessions which of course govc does
not know about. This will be added in a similar fashion.

This code is ultimately necessary to ensure acceptance tests work, which
create enough new sessions to hit the default 1000 session limit in
vSphere, mainly from the re-creation of the provider every time
testAccProviders() is called. Note that just fixing this helper won't
work, as the test framework we currently use for the acceptance tests
runs individual tests in their own sub-process as well, meaning that
there would not be any re-use across the entire set of tests, just
within the one test, so improvements would be marginal.