provider: Update tags client, fix some log messages
We were using the incorrect method type for checking on sessions in the
work we are doing in the VIC tags client. Fixed that now, and session
persistence for tags is now working.

Also corrected/added a couple of log messages for the REST client to
make the REST client configuration process more/properly verbose.