provider: Fix keys in configuration function, refactor stack
This should have been tested properly in #422, but the way
providerConfigure was returning a fully configured Client versus the
configuration made it hard to do so, so it was skipped, and of course I
was burned as a result. :P

This fixes things so that the actual creation of the Config struct is
moved to a separate NewConfig function, which allows us to test for the
valid config keys. providerConfigure then just calls that.