r/virtual_machine: Fix broken acceptance tests and CD drives
This does a few things:

* Looks like some acceptance test were broken - first one was due to
something changing in RancherOS that does not allow for the proper
hot-remove of virtual NICs anymore. I've adjusted this test so it
powers off the VM first, which is not so much an issue as we aren't
testing for a hot-remove anyway, but more to make sure that things are
consistent in state.
* The vApp property tests were missing settings for thin_provisioned and
eagerly_scrub, which should be coming from the template data sources.
These were failing as a result as the CoreOS OVA does not use thin
provisioned disks.
* Finally, #421 introduced a change where we were actually failing if we
didn't recognize the CD drive type. There are a few we still don't
support, namely passthrough CD drives. I've changed this to a debug log
message, and a unsetting of all parameters. This should ensure that
devices that aren't recognized but need to be changed to a recognized
device are properly accounted for in the diff, in addition to removing
unrecognized devices in the event no CDROM drive is necessary.