r/vmfs_datastore: Add datastore cluster support
This adds the datastore_cluster_id attribute to the VMFS datastore
resource, allowing one to add a VMFS datastore resource to a datastore
cluster.

In addition to this, a bit of refactoring was done to remove some unused
error messages, and adding additional disks during creation was moved up
in precedence so that the ID could be set earlier, removing the need for
rollback on folder moves and ID/custom attribute modification.