r/virtual_machine: Add datastore_cluster_id attribute to schema
This is in prep for adding datastore cluster support.

We may need extra validation aside from what is listed here, but this
gives us at least a bit of a baseline to work from - illustrating where
the conflicting attributes lay as well so that config validation versus
diff validation can block the more basic cases.