Capture updated device list when destroying disks
The procedure "applyDeviceChange" modifies the
object.VirtualDeviceList it is passed, and returns an updated
reference to the object. As an object.VirtualDeviceList is a slice,
and modifying elements of a slice modifies the underlying array,
failing to capture this return value results in an inaccurate view of
the available devices. In this case, when destroying a disk, if the
last element of the object.VirtualDeviceList is a disk,
"findVirtualDiskByUUIDOrAddress" will incorrectly report that there
are multiple disks with the same UUID.

Fixes issue #404.