r/virtual_machine: Much cleaner storage DRS VM creation path
This is a new attempt at VM creation through storage DRS. This time,
unlike the last attempt, we use the storage DRS API completely as a
replacement to CreateVM_Task. This is how it's supposed to be intended.
The PoC in govc does not do this, rather it assigns the datastores in
the returned recommendations to the VirtualMachineConfigSpec manually.
In our provider, this resulted in a lot of convoluted code and workflows
that just were not worth it in the long run, especially for cloning.

The workhorse for this are helpers for both datastore cluster and
standard datastore creation paths, in addition to a new CreateVM helper
in the storagepod helper package that is nearly API compatible with our
Create function in the virtualmachine helper package, with the only
addition being that we require the datastore cluster to deploy to.

We will use this PoC for VM creation to create helpers for cloning,
reconfiguration, and migration as well. All of these should be similarly
simple to creation.

The only drawback to this approach is that immediate reconfiguration of
the VM after cloning to do things like grow the disk or what not may
result in SDRS recommendations that may either cause the operation to
fail or result in an immediate storage vMotion should the
reconfiguration operation push the disk into SDRS violation territory.
However, when comparing how complicated the path of getting
recommendations for a new VM with the final specs and transposing those
recommendations on the clone spec, versus just making the two SDRS apply
calls, the latter is much cleaner and less complicated from a code
perspective - the former would require massive refactoring of the
current resource to be accommodated with any level of sanity.

TODO: Add datastore cluster self-hosting into acceptance tests.