r/virtual_machine: Fully support SDRS for cloning
This makes the few minor changes necessary to the cloning workflow to
ensure that a datastore ID does not need to be specified. It also adds a
step to the reconfigure operation that just defers to standard
reconfigure if there are actually no disk creation operations happening
in a reconfigure, as an error is issued if there are no SDRS disk
placements requested.