r/virtual_machine: Storage DRS support for VM migration
This commit includes pretty much complete support for vMotion through
the Storage DRS API, which we use if we have a datastore cluster
defined.

The workflow is properly triggered not necessarily by checking if there
is a change in datastore_cluster_id, but rather checking the datastores
that VM configuration and individual disks are in, and ensuring that
those datastores are members of the defined datastore cluster. We also
handle the event in which a datastore cluster ID is not known but
defined at diff time (ie: pending interpolation), which basically will
force the default datastore back to computed, which will in turn request
a vMotion.

Some extra tests are/may be pending:

* (Only) host migration when a datstore cluster is defined. This will
determine if we need to handle the case where it's possible that no SDRS
recommendations will be given.
* Migration of an individual disk to a datastore that is not a member of
the defined datastore cluster. This is to determine if the resource will
properly trigger a migration to move the disk back.
* Possible handling of a scenario where TF has triggered a change in
datastore cluster (ie: set it back to computed) but no real change is
happening (ie: the possibly the current cluster is being destroyed and
its datastores moved to a newly created cluster). Admittedly this is an
edge case, and depending on the difficultly of writing the test (which
should be an indication of the likelihood of the edge case happening),
we may just skip this. Technically, this poses the possibility of a
failed operation due to no recommendations, but would more than likely
resolve itself in the next TF plan, which would then be empty on part of
the new datastore cluster ID being available, with the datastore the VM
resides in being a member.