r/virtual_machine: More tests for Storage DRS vMotion
As mentioned in the previous commit, this update adds two out of the
three mentioned tests: the host vMotion test, and the individual disk
migration test. We found a minor bug in the latter path that was fixed.

The last path I have opted to not add. The main reason being is that
it's too hard to predict that we would not have a race condition under
this scenario in normal use without even more use of workarounds like
depends_on, as more than likely any changes that would happen within
datastore clusters would be happening in the graph the same time as the
virtual machine changes would be happening, which could possibly lead to
SDRS errors for more standard reasons. The best solution would be to
manage these things in separate stacks anyway, and the likelihood that
people are going to be managing datastore clusters and VMs in the same
configuration is probably small anyway.