r/storage_drs_vm_config: Tests
Adding what should be functional tests - functional in that they will
run, not that they will pass. ;)

Also, removed the default for sdrs_automation_level. This is a string
and not a string pointer, as such it was assumed that this could not be
ommitted, but it is tagged as omitempty. Omitting it from the config
should ensure it uses the cluster setting.