r/storage_drs_vm_config: Add import/update tests, save IDs in Read
Saving datastore cluster and VM IDs in read now as well to add
completeness to the import process.

Update seems to be broken, some nil values do not seem to be saving and
we need to figure out why that is. Depending on the outcome we might
just ForceNew all attributes versus update.