r/virtual_machine: Streamline disk relocate check
This updates the vMotion check so that a change in datastore cluster ID
is handled properly.

The old logic included a change check to datastore_cluster_id, only
really to ensure that we caught any possible changes to disk
sub-resources when a datastore cluster was defined (ie: in the event of
relocating a disk that had been migrated out of the cluster). This was
producing a side-effect that relocations where being attempted even when
there was no actual changes in datastores.

The new logic adds a return flag to the DiskRelocateOperation function
for disk sub-resources, which gets flagged on the first genuine
relocation operation that gets processed - including default datastore,
pinned datastore, and datastore cluster-triggered changes. This allows
us to check for these things without triggering a migration due to a
change in datastore_cluster_id that does not result in any actual
change.

Removing datastore_cluster_id from the logic does not have any
implications to other migration workflows as all of these operations are
handled naturally through diff customization. The source of truth for
whether or not a migration is necessary are always the datastore_id
attributes.