r/storage_drs_vm_config: Fix 3-state bool diffs
This fixes the systemic issues with dealing with TF 3-state bools and
working with diffs in Terraform core.

There are currently a couple of issues here:

* GetOkExists does not function correctly when there are values in the
state. If a value enters state it "exists", even when a value is being
removed in config.
* Diffs from zero values to nils do not register. This is because the
diff code in the SDK was not designed with consideration for nils being
a valid state - nils are converted to zero values and then ignored when
they are compared, as the value will then be equal to the zero value
that exists in state.

We fix this for bools by using string fields, and then normalizing and
validating all possible values to 3 values that will always match -
"true", "false", and an empty string. This exploits current HCL
functionality as well where mapstructure converts bareword bools to
string values when the field type is TypeString.

Note that this is not an issue that just applies to this resource. It's
probably prudent to go back through some of the other resources (namely
standard and distributed port groups) and check to see if we need to fix
these things there as well.

This also may need to be fixed for integer values as well, as all 3 of
these issues apply there as well (setting nil will save as 0, going from
0 => nil will not trigger a diff). The issue with GetOkExists applies to
*all* Terraform things and is a larger issue, even with strings.