r/virtual_machine: Backings should never be added to disks w/SDRS
The current disk sub-resource code only skips adding a backing when a
datastore_id is absent. This is only true when using SDRS on VM
creation.  On update, since a datastore has been populated in state from
a refresh call, datastore_id will be populated, producing the backing in
the VirtualMachineConfigSpec. In SDRS this is actually how you specify
that you don't want SDRS to recommend a datastore to place the disk in,
but rather want the disk to go into a specific datastore. This is not
what we want.

The tests included with this PR are mainly for sanity checking for
SDRS-enabled VMs with multiple disks, and ensuring that update
operations work correctly. To actually test for the presence of a nil
backing in the above described scenario, one needs to use delve from the
make testacc target:

make testacc TEST=./vsphere \
  TESTARGS=-test.run=TestAccResourceVSphereVirtualMachine_datastoreClusterCreateUpdate

Set a breakpoint in vsphere/internal/helper/storagepod in the
ReconfigureVM function, which will break on the reconfigure step. You
can then check
spec.DeviceChange[0].Device.Backing.VirtualDeviceFileBackingInfo.