r/virtual_machine: Fix vApp ISO tests
This fixes the vApp ISO tests, which were failing because the Ubuntu
OVAs do not have a way of getting a static network configuration, hence
they wait 5 minutes on DHCP on our static network.

We had an image we hacked to lower this timeout, but the template was
recently re-uploaded, and the hacking process is cumbersome (involving
mounting the VMDK in the OVA to add an override to networking.service),
and is not easily repeatable with our current process.

As the acceptance tests are automated, waiting the extra time for the
DHCP timeout is not a big deal. Hence we've just upped
wait_for_guest_net_timeout to 10 minutes on the appropriate fixtures.