r/virtual_machine: Remove clone config restriction for import/legacy
This removes the restriction on the presence of the clone sub-resource
block for imported and legacy resources.

Previously, this was blocked as it forced a new resource, which is
obviously not what people would want during this process. ForceNew was
(naturally) controlled in schema, so there was not much we could do
here.

Some recent changes to diff customization has allowed us to change this
process in a way that has been relatively non-invasive. This has
resulted in the changes below:

* Rather than using schema to force new resources for clone attributes,
we now go through changed keys in the clone namespace and ForceNew these
keys through diff customization. This happens for both new and existing
resources to prevent diff mismatches.
* We leverage imported as a conditional to provide a one-time exemption
from this behavior. We now also transition imported from true to false
after the first apply post-import/migration.

These two changes have the effect in that on the first apply operation
after a resource has been imported or transitioned from a legacy version
of the provider, all clone configuration settings are simply persisted
to state, without forcing a new resource. This, like other post-import
operations like flagging keep_on_remove for disks, is a no-op on the
vSphere API side and does not trigger a reconfigure or a rebuild of the
VM.

The new behavior has been documented. It should be noted that if
configuration changes to clone are made in the first post-apply run they
will be ignored permanently as well (as they will persist to state
without any action carried out on them), however this should be
avoidable when import/migration instructions are followed. This has been
documented as well.