d/virtual_machine: Nil check for empty VM configuration
This handles a crash case when there is no configuration for a specified
virtual machine.

This unfortunately is an edge case and hard for us to test for, as cases
like this where the VM lack a config attribute are usually due to issues
with importing the VM, or an orphaned virtual machine in vSphere.