r/virtual_machine: Set datastore_id in disk sub-resource to computed
This is something that has been an issue for while but is now even more
prominent now that datastores don't necessarily need to be specified at
all. Owing in part that datastore_id is not populated during the initial
TF apply when using the default datastore for disks, changing both the
default datastore (or now, possibly, datastore clusters) and a key that
forces a new resource at the same time will create a diff mismatch.

Setting datastore_id in the disk sub-resource schema fixes this issue.