r/virtual_machine: Controllable routable network waiter behavior
This adds the wait_for_guest_net_routable setting, which allows for the
modification of the behavior of the network waiter with relation to
whether or not it compares IP addresses it discovers against the default
gateway.

When false, the waiter will ignore any gateways - it will exit
successfully on discovery of any "real" address - in other words, one
that is not link-local/loopback/multicast, regardless of if the address
does not match a discovered default gateway, or regardless if a gateway
has even been discovered yet.