helper/dvportgroup: Handle nil results from FromKey
I'm not too sure how this can happen exactly, but we have gotten a crash
report on it (#456). Every invocation of DVSManagerLookupDvPortGroup
that I can see should return some sort of fault, however the report here
indicates that under some scenarios, a nil result can be returned
instead with no fault.

This patch handles this, wrapping the case in a specific error. Based on
the results of this (we will probably have to rely on user feedback to
see what exactly kinds of scenarios this can happen under), we might
handle the wrapped error higher up the stack and possibly ignore it,
depending on whether or not this is coming from cloned templates
immediately on post-refresh.