r/virtual_machine: Different approach to datastore_id diff mismatch
After discussing the change to datastore_id in disk sub-resources to
computed, it was determined that there could possibly be scenarios where
this would be damaging to a diff where someone was creating a new disk
at a list location that contained a previous disk's state. This would
result in the old datastore information being grafted into the new disk
due to how Optional/Computed works.

This reverts the schema to the old non-computed form, opting instead to
replace a blank datastore_id with "<computed>" in disk diff when
datastore_id is not supplied. This is then treated the same as an empty
value on virtual disk creation.