r/virtual_machine: Allow vMotions to other clusters
As host_system_id is an optional attribute that also holds the current
host that a VM is currently located on (read: Optional/Computed value),
changing resource_pool_id to a resource pool in a cluster that does not
contain the host denoted by the current value of host_system_id (which
would be unmodified unless actually pinned by the user) would produce an
error.

This corrects this by flagging host_system_id as computed in the diff if
resource_pool_id is changing while the former is not. This correctly
skips the host check.