r/virtual_machine: Restore ID availability to rollback functionality
The ID of the resource is actually necessary for rollback as we send the
VM through full resource deletion during rollback (this is to safely
detach any disks that may have been attached already). This commit moves
the ID step back to where it previously was. However, we clear the ID no
matter what in the rollback function (via a defer at the start of the
rollback function).