r/compute_cluster: Resource finished
Finished rest of CRUD and import, resource is now ready to have
acceptance tests written for it.

Commit also contains some small modifications to the
vsphere_datastore_cluster resource as I noticed some error messages were
off as I was copying the code (a lot of the code was drawn from this
resource as a template).