r/compute_cluster: Add VM test
Forgot to add this one - does a test where we create a VM in the
cluster.

This one depends on the test environment having the same setup as the
live cluster normally used for VM testing. This is the case in our own
lab and was how VSPHERE_ESXI_HOST5-7 were intended to be used.