r/compute_cluster: Extra import support
Added a function to set defaults for all attributes that are not covered
in read, except for datacenter_id which we do in a separate read
function to make sure that is always set, even on import. This allows a
clean import test without requiring any variables be entered into
ImportStateVerify.