d/compute_cluster: New data source
This adds the vsphere_compute_cluster data source, which allows one to
look up an existing cluster.

It also closes a convenience gap in the provider which has been a point
of confusion since around 1.0, in that people looking to supply the
resource_pool_id for a virtual machine were confused in that they didn't
think they had a resource pool. This now allows someone to specify a
cluster directly, and then use the resource_pool_id attribute from this
data source as the input to vsphere_virtual_machine.