r/compute_cluster_vm_group: Rename, tests, and fixes
A number of things are in this update:

* First off, we've renamed the resource to
vsphere_compute_cluster_vm_group.  This is because the groups are
actually located within a cluster configuration and are not
transferrable to any other kind of cluster (ie: datastore clusters,
which actually don't have groups at all).
* Second off, we've had to add extra lookup logic to translate VM UUIDs
to MoRefs. This has been added into the shared helper package so it can
be re-used for upcoming work easier (as we add rules, there will be a
number of places this needs to be changed too).
* Lastly, tests have been added and the work necessary to get them
working has been done.

Completion is pending documentation.