r/compute_cluster_vm_host_rule: Documentation
Complete docs for the vsphere_compute_cluster_vm_host_rule resource.

There is a bit of doc work that is necessary in other resources that
will be following in another commit - it looks like in the work that has
been done recently, a lot of which has been sourced from existing docs -
some things weren't necessarily changed that should be. That will be
following next commit.