Fall back to guest ipAddress
Older versions of VMTools do not populate the IpStack and Net propreties
in GuestInfo which is what is currently used to find the VM's IP
address. With this change we now have the option to wait for the IpAddress
property instead which does get populated with older versions of
VMTools.

In order to ensure that we wait for the guest IpAddress property to be
populated, a new wait_for_guest_ip_timeout configuration option has been
added. This option is disabled by default but can be enabled by setting
the timeout to the number of minutes to wait for the guest IpAddress
property to be populated.