Add Subcategory for Resources
Add subcategories to docs for display on the Terraform Registry

We will soon be displaying documentation for this provider both on
[terraform.io](https://www.terraform.io/docs/providers/index.html)
and on [the Terraform Registry](https://registry.terraform.io/providers).

Documentation displayed on the Terraform Registry will not use the ERB
layout containing the existing navigation, and will instead build the
navigation dynamically based on the directory and YAML frontmatter of a file.
For providers that group similar resources and data sources by service or use
case (subcategories), we'll need to add this information to the Markdown
source file.