Wait until VM is ready to be powered on
Some times a virtual machine is not available to be powered on
immidiately after it gets created. To this end each vSphere object
has a list of disabled methods that cannot be called.

This change introduces a function that blocks until PowerOnVM_Task
call is no longer in the list of disabled methods.