Do not throw error when disk path is not known yet
If the path to an attached disk is interpolated, then sometimes we may
end up in a situation where the diff for the VM having the disks runs
before the disk itself has been processed. In this case since the path
has no value we throw an error.

This changes the behaviour so we check if the path has been calculated
before we consider it final or not.