vsphere_virtual_machine: fix broken codeblock
Formatting is broken on the online docs: https://www.terraform.io/docs/providers/vsphere/r/virtual_machine.html#deploying-vm-from-an-ovf-template
The markdown codeblock is not rendered correctly. It works fine on GitHub. Perhaps the CSS for the hashicorp docs site requires a newline before codeblocks, as is implemented by the other code blocks.