Create `vsphere_ovf_vm_template` datasource
The provider expects to be the only entity altering the settings of a
virtual machine, which is not the case when deploying via OVF.

The easy way out of this would be to make all attributes `Computed`
which would make terraform to blidnly accept all changes into the state.
Unfortunately this would also mean a loss of drift detection.

The middle ground is to create a new datasource that extracts all the
settings from the OVF template and expose it as outputs, so we can then
pass it as attribute values to the `vsphere_virtual_machine` resource.

The ideal solution would be to create a whole new resource to deal
exclusively with OVF-built virtual machines but it would be a breaking
change since users are already using the ovf subresource in
`vsphere_virtual_machine`. This could potentially go into a v2.0 version
of the provider (cc: @redeux).

This commit creates the `vsphere_ovf_vm_template` resource and re-arranges
some of the OVF code so it can be used from both `vsphere_virtual_machine`
resource and the datasource.